9/13 9:56pm Majorly cleaned up the code.
Changed the id's of the table headers in activity_main.xml

added no-arg constructor in Album -- app crashes w/o one
firebase also requires POJO identifiers to be public

added getView() to our AlbumArrayAdapter -- this is what changes our listView, that is CURRENTLY in activity_main.xml

In MainActivity:

in onCreate() (on startup) I've added a get() using our database to read all the current documents that are in the database:
we are using an arrayList of type Albums and firestore uses our Album class to create Album objects with the fields from
our database. Here is where our issue lies.

In line 56 when we call document.toObject(Album.class) each Album object's field (artist, year, etc.) is NULL
therefore when the list of albums are printed to the table it prints blank fields. Android Studio is giving warnings about
@NonNull and @Nullable that I have not been able to dig into yet. Could help fix this.

And that's where we are...